Refactor: Main Component
