+++ src/pages/Dashboard.jsx
@@ -1,7 +1,7 @@
 import React, {Component} from 'react';
 
 // import Header from '../components/Header';
-import Main1 from '../components/Main1';
+import Main from '../components/Main';
 import NavBar from '../components/NavBar';
 
 export default class Dashboard extends Component {
@@ -9,7 +9,7 @@ export default class Dashboard extends Component {
         return (
             <div className="wrapper">
                 <NavBar show/>
-                <Main1/>
+                <Main/>
             </div>
         )
     }
